% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisquare.R
\name{bisquare}
\alias{bisquare}
\title{Calculate weights via the bisquare kernel}
\usage{
bisquare(d, bw)
}
\arguments{
\item{d}{vector of distances from the center of the bisquare kernel}

\item{bw}{bandwidth parameter for the bisquare kernel}
}
\value{
The weight given to each of the distances \code{d} by a bisquare function with the given bandwidth \code{bw},
which is zero if \code{d>bw} and  \code{(1-(d/bw)^2)^2} otherwise
}
\description{
Calculate weights via the bisquare kernel
}

