% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjugate.gradient.R
\name{conjugate.gradient}
\alias{conjugate.gradient}
\title{Conjugate gradient optimization}
\usage{
conjugate.gradient(objective, gradient, y, X, S, beta, wt, ltau, M, logV,
  verbose = TRUE, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{objective}{function to be minimized (i.e. the marginal likelihood)}

\item{gradient}{function to calculate the gradient of the objective}

\item{y}{response variable for the regression function}

\item{X}{design matrix of covariates for the systematic portion of the model}

\item{S}{spatial random effect design matrix for the stochastic portion of the model}

\item{beta}{vector of initial regression coefficients for the fixed effects}

\item{wt}{vector giving the weight of each observation}

\item{ltau}{initial value for logarithm of prior distribution on the precision of the random effect loadings, \code{u} (precision component)}

\item{M}{vector of assumed posterior means of the spatial random effects}

\item{logV}{initial value of logarithm of the entries of the random effect posterior covariance matrix}

\item{verbose}{indicates whether to write detailed progress reports to standard output}

\item{tol}{proportional change in likelihood smaller than \code{tol} indicates convergence}
}
\value{
List consisting of \code{par}, the value of parameters that minimize
the objective function, and \code{value}, the value of the minimized objective
}
\description{
Maximize the marginal log-likelihood with respect to the log covariance matrix of the spatial random effects.
}
\details{
Uses the method of conjugate gradient to maximize the marginal log-likelihood of the Cox process with respect to the log covariance matrix of the random effects. Uses backtracking to determine the step size, reducing the step size by half until stepping causes a decrease in the deviance. In the past I had used majorization-minimization to estimate an optimal step size, but the current approach seems to converge more quickly. The maximum number of steps before restarting conjugacy is the number of free parameters. In this case, that means the number of elements in the upper triangle (including the diagonal), which is r-choose-2 (where r is the dimension of the covariance matrix). Use only the upper diagonal because the covariance matrix because the covariance matrix is symmetric. Account for symmetry by doubling the gradient for all off-diagonal entries. Final convergence is when the relative decrease in the deviance is less than \code{tol}.
}

