% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.laplace.R
\name{cox.laplace}
\alias{cox.laplace}
\title{Estimate the regression parameters of a Cox process model using the Laplace approximation.}
\usage{
cox.laplace(y, X, S, wt, beta.start, tol = sqrt(.Machine$double.eps),
  verbose = TRUE)
}
\arguments{
\item{y}{vector of response data}

\item{X}{design matrix for fized effects}

\item{S}{design matrix for the spatial random effects}

\item{wt}{vector of observation weights}

\item{beta.start}{initial values of the fixed effect coefficients (used for iterative estimation scheme)}

\item{tol}{tolerance for judging convergence of the algorithm}

\item{verbose}{if \code{TRUE}, the algorithm prints verbose updates on its progess}
}
\description{
\code{cox.laplace} uses a Laplace approximation to integrate spatial random effects out of a Cox process regression model.
The Laplace approximation works by maximizing the joint likelihood of the data and random effects, then approximating the
random effects as multivariate normal via a second-order Taylor expansion.
}

