% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.variational.R
\name{cox.variational}
\alias{cox.variational}
\title{Estimate the regression parameters of a Cox process model using the variational approximation with unconstrained covariance for the random effects.}
\usage{
cox.variational(y, X, S, wt, beta.start, tol = sqrt(.Machine$double.eps),
  verbose = TRUE)
}
\arguments{
\item{y}{vector of response data}

\item{X}{design matrix for fized effects}

\item{S}{design matrix for the spatial random effects}

\item{wt}{vector of observation weights}

\item{beta.start}{initial values of the fixed effect coefficients (used for iterative estimation scheme)}

\item{tol}{tolerance for judging convergence of the algorithm}

\item{verbose}{if \code{TRUE}, the algorithm prints verbose updates on its progess}
}
\description{
\code{cox.variational} uses a variational approximation to estimate the parameters of a Cox process regression model with spatial random effects.
For this function, the variational approximation to the posterior distribution of the spatial random effects is a multivariate normal with general
covariance matrix.
}

