\name{ForestPlotCox}
\alias{ForestPlotCox}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
ForestPlotCox
}
\description{
Make a forestplot visualization of the cox regression.
}
\usage{
ForestPlotCox(models)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{models}{
Output of the CoxSummary function (summ=TRUE, styled= FALSE)
}
 \item{...}{
See the documentation of the foresplot function.
}
}
\references{
https://cran.r-project.org/web/packages/forestplot/index.html
}
\author{
Tamas Szili-Torok
}

\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
ForestPlotCox (models)
{
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
