% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.variational.independent.R
\name{likelihood.bound.indep}
\alias{likelihood.bound.indep}
\title{Variatonal lower bound on likelihood (eq 3.1 of Ormerod and Wand, 2012)}
\usage{
likelihood.bound.indep(M, log.diagV, ltau, y, X, S, beta, wt)
}
\description{
Variatonal lower bound on likelihood (eq 3.1 of Ormerod and Wand, 2012)
}

