% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.variational.R
\name{likelihood.bound}
\alias{likelihood.bound}
\title{Variatonal lower bound on likelihood (eq 3.1 of Ormerod and Wand, 2012)}
\usage{
likelihood.bound(y, X, S, beta, wt, ltau, M, V)
}
\description{
Variatonal lower bound on likelihood (eq 3.1 of Ormerod and Wand, 2012)
}

