% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.laplace.R
\name{likelihood.joint}
\alias{likelihood.joint}
\title{Calculate the (negative) marginal log-likelihod}
\usage{
likelihood.joint(par, y, X, S, wt, tol, verbose)
}
\arguments{
\item{par}{vector of fixed-effect coefficients with log-precision parameter of the random effects appended}

\item{y}{vector of Poisson-distributed responses for DWPR}

\item{X}{matrix of fixed-effect covariates}

\item{S}{matrix of random-effect covariates}

\item{wt}{vector of observation weights}

\item{tol}{relative change threshold for convergence}

\item{verbose}{output detailed information about the progress of the algorithm?}
}
\description{
Calculate the (negative) marginal log-likelihod
}

