% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.laplace.R
\name{likelihood.laplace}
\alias{likelihood.laplace}
\title{Calculate the (negative) marginal log-likelihod}
\usage{
likelihood.laplace(beta, ltau, y, X, S, wt, tol, verbose)
}
\arguments{
\item{beta}{vector of fixed-effect coefficients}

\item{ltau}{log-precision of the random effects}

\item{y}{vector of Poisson-distributed responses for DWPR}

\item{X}{matrix of fixed-effect covariates}

\item{S}{matrix of random-effect covariates}

\item{wt}{vector of observation weights}

\item{tol}{relative change threshold for convergence}

\item{verbose}{output detailed information about the progress of the algorithm?}
}
\description{
Calculate the (negative) marginal log-likelihod
}

