\name{NumToString}
\alias{NumToString}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
NumToString
}
\description{
This function takes a number with n decimal places and converts it to a string with desired decimal places.
}
\usage{
NumToString(x, dec)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Numerical which you would like to convert to a string.
}
\item{dec}{
Desired decimal places of the output.
}
}
\author{
Tamas Szili-Torok
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
NumToString (8.98699,2)
{
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
