\name{PlotHr}
\alias{PlotHr}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Plot Hazard ratio
}
\description{
Plot the hazard ratio of Cox propotional hazard regression
}
\usage{
PlotHr(prediction, df, ind.var)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{pred}{
Prediction
}
\item{df}{
DataFrame
}
\item{ind.var}{
Name of independent variable in string format
}
}

\references{
%% ~put references to the literature/web site here ~
}
\author{
Tamas Szili-Torok
}

\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x)
{
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
