\name{PVal}
\alias{PVal}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
PVal
}
\description{
Convert a low P value to a string of '<0.001'.
}
\usage{
PVal(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A numerical.
}
}

\author{
Tamas Szili-Torok
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
PVal (0.000089)
{
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
