% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.variational.independent.R
\name{score.diagV}
\alias{score.diagV}
\title{Gradient of the lower likelihood bound w.r.t. V:}
\usage{
score.diagV(logV, M, ltau, y, X, S, beta, wt)
}
\description{
Gradient of the lower likelihood bound w.r.t. V:
}

