% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.variational.independent.R
\name{score.fin.indep}
\alias{score.fin.indep}
\title{Gradient of the lower likelihood bound w.r.t. V:}
\usage{
score.fin.indep(par, y, X, S, wt)
}
\description{
Gradient of the lower likelihood bound w.r.t. V:
}

