% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.variational.R
\name{score.fin}
\alias{score.fin}
\title{Gradient of the variational negative log-likelihood with respect to beta, M, and ltau}
\usage{
score.fin(par, y, X, S, wt, V)
}
\arguments{
\item{par}{vector composed of \code{beta}, \code{M}, and \code{ltau}}

\item{y}{vector of response values}

\item{X}{matrix of fixed effect covariates}

\item{S}{design matrix for the spatial random effects}

\item{wt}{vector of observation weights}

\item{V}{converged covariance matrix for the random effects, estimated by a variational approximation}
}
\value{
vector of gradients of the variational negative log-likelihood with respect to \code{par}
}
\description{
Gives the gradient of the variational negative log-likelihood with respect to coefficients \code{beta}, precision component \code{ltau}, and mean vector of the variational approximation, \code{M}. Within the \pkg{cox} package, this function is used to calculate gradients within a call to \code{optim} after the variational approximation has converged, as part of the process of calculating the Hessian.
}

