% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.laplace.R
\name{score.laplace}
\alias{score.laplace}
\title{Gradient of the marginal log-likelihood for a Cox process}
\usage{
score.laplace(beta, ltau, y, X, S, wt, tol, verbose)
}
\arguments{
\item{beta}{fixed-effect coefficient vector}

\item{ltau}{log precision of the random effect coefficients}

\item{y}{vector of Poisson-distributed responses for DWPR}

\item{X}{matrix of fixed-effect covariates}

\item{S}{matrix of random-effect covariates}

\item{wt}{vector of observation weights}

\item{tol}{relative change threshold for convergence}

\item{verbose}{output detailed information about the progress of the algorithm?}
}
\description{
Calculate the gradient of the marginal log-likelihood with respect to coefficients beta and log variance component ltau. Uses the Laplace approximation.
}

