% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.variational.R
\name{score.logV}
\alias{score.logV}
\title{Gradient of the variational likelihood bound w.r.t. log of covariance matrix}
\usage{
score.logV(logV, y, X, S, beta, wt, ltau, M)
}
\arguments{
\item{logV}{logarithm of vectorized upper triangle of the covariance matrix of random-effect coefficients for the variational approximation}

\item{y}{vector of Poisson-distributed responses for DWPR}

\item{X}{matrix of fixed-effect covariates}

\item{S}{matrix of random-effect covariates}

\item{beta}{fixed-effect coefficient vector}

\item{wt}{vector of observation weights}

\item{ltau}{log precision of the random effect coefficients}

\item{M}{vector of means of random-effect coefficients for the variational approximation}
}
\description{
Gradient of the variational likelihood bound w.r.t. log of covariance matrix
}

