% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim.cox}
\alias{sim.cox}
\title{Simulate a Gaussian random field a set of points}
\usage{
sim.cox(intercept, beta, rando.corr, rho, n.q, n, output.loc, reps)
}
\arguments{
\item{intercept}{intercept for the regression portion of the model}

\item{beta}{coefficients of the regression portion of the model}

\item{rho}{correlation between the}

\item{n.q}{number of quadrature points}

\item{output.loc}{save output to this folder}

\item{reps}{number of times to repeat this simulation setting}
}
\description{
Simulate a Gaussian random field a set of points
}

