% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjugate.gradient.R
\name{conjugate.gradient}
\alias{conjugate.gradient}
\title{Maximize the marginal log likelihood with respect to the covariance matrix of the spatial random effects}
\usage{
conjugate.gradient(objective, gradient, y, X, S, beta, wt, ltau, M, logV,
  verbose = TRUE, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{objective}{function to be minimized (i.e. the marginal likelihood)}

\item{gradient}{function to calculate the gradient of the objective}

\item{y}{response variable for the regression function}

\item{X}{design matrix of covariates for the systematic portion of the model}

\item{S}{spatial random effect design matrix for the stochastic portion of the model}

\item{beta}{vector of initial regression coefficients for the fixed effects}

\item{wt}{vector giving the weight of each observation}

\item{ltau}{initial value for logarithm of prior distribution on the precision of the random effect loadings, \code{u} (precision component)}

\item{M}{vector of assumed posterior means of the spatial random effects}

\item{logV}{initial value of logarithm of the entries of the random effect posterior covariance matrix}

\item{verbose}{indicates whether to write detailed progress reports to standard output}

\item{tol}{proportional change in likelihood smaller than \code{tol} indicates convergence}
}
\value{
List consisting of \code{par}, the value of parameters that minimize
the objective function, and \code{value}, the value of the minimized objective
}
\description{
Maximize the marginal log likelihood with respect to the covariance matrix of the spatial random effects
}

