% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRE.R
\name{SRE}
\alias{SRE}
\title{Set up the multiresolution spatial basis functions}
\usage{
SRE(loc, n.1 = 16, n.res = 3, bw.scale = 1.5, res.scale = 3,
  xlim = NULL, ylim = NULL)
}
\arguments{
\item{loc}{matrix of locations (both point observations and quadrature locations) that are to be included in the Cox process model. The columns must be named \code{x} and \code{y}.}

\item{n.1}{number of basis functions that make up the first (coarsest) resolution}

\item{n.res}{number of resolutions to use}

\item{bw.scale}{multiplier used to set the bandwidths of the basis functions. Bandwidth for the \code{k}th resolution is set to \code{bw.scale * min.dist}, where \code{min.dist} is the minimum distance between basis function centers in the \code{k}th resolution (default is 1.5).}

\item{res.scale}{relative increase in the number of basis functions to use for each resolution (default is 3)}

\item{xlim}{bounds (in the x dimension) of the region to be spanned by the basis functions (optional)}

\item{ylim}{bounds (in the y dimension) of the region to be spanned by the basis functions (optional)}
}
\value{
\code{SRE}, the spatial random effect design matrix, with a row for each location and a column for each basis function. The \code{(i,j)}th entry of this matrix is the value of basis function \code{j} at location \code{i}.
}
\description{
Set up the multiresolution spatial basis functions
}

